Add backup restoration to Gotify chart
Add backup restoration using an init container to the Gotify chart. When
using an init container, the service's container will wait for the
backup restore to finish before going up.
This became problematic with Gitlab when it's backup grew beyond a
couple of GBs, because I've been blindly waiting for restoration up
until now and Gitlab really shouldn't boot before the backup is fully
restored.